refactor: extract logic and state to hook useAllowedUsers
