+++ src/pages/playlist/[id].tsx
@@ -1,17 +1,12 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { GetServerSideProps, NextPage } from 'next';
-import { CookieKey, Playlist, QueryKey, UserProfile } from '../../@types';
+import { CookieKey, Playlist, UserProfile } from '../../@types';
 import { parseCookies } from 'nookies';
-import {
-    getPlaylist,
-    getUserInfo,
-    getUserProfile,
-    getUserProfiles,
-    updatePlaylistAllowedUsers,
-} from '../../services/api';
-import { useMutation, useQuery } from '@tanstack/react-query';
+import { getPlaylist, getUserInfo, getUserProfile } from '../../services/api';
 import { UsersSearchBox } from '../../components/UsersSearchBox';
 import Link from 'next/link';
+import { useAllowedUsers } from './hooks/useAllowedUsers';
+import Image from 'next/image';
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
     const tokenCookieKey: CookieKey = 's-p-guard:token';
@@ -73,123 +68,13 @@ const Playlist: NextPage<PlaylistProps> = ({
     allowedUsers,
     ownerSpotifyId,
 }) => {
-    const [users, setUsers] = useState<
-        {
-            id: string;
-            name: string;
-            imageURL: string;
-            status: 'permanent' | 'idle' | 'removed' | 'added';
-        }[]
-    >(() =>
-        allowedUsers.map(({ id, name, image_url }) => ({
-            id,
-            name,
-            imageURL: image_url,
-            status: id === ownerSpotifyId ? 'permanent' : 'idle',
-        })),
-    );
-    const usersProfilesKey: QueryKey = 'users-profiles';
-    const usersProfilesQuery = useQuery([usersProfilesKey], {
-        queryFn: () => {
-            return getUserProfiles(
-                users
-                    .filter(({ status }) => status !== 'removed')
-                    .map(({ id }) => id),
-            );
-        },
-        initialData: allowedUsers,
-        keepPreviousData: true,
-        staleTime: Infinity,
-    });
-    const usersProfilesMutation = useMutation({
-        mutationFn: async (userIds: string[]) =>
-            updatePlaylistAllowedUsers(playlist.id, userIds),
-        onSuccess: () => {
-            usersProfilesQuery.refetch();
-        },
-    });
-
-    const addNewAllowedUser = (
-        newUser: Omit<typeof users[number], 'status'>,
-    ) => {
-        setUsers((state) => {
-            return Array.from(
-                new Set([...state, { ...newUser, status: 'added' as const }]),
-            );
-        });
-    };
-
-    const removeNewAllowedUser = (state: typeof users, idPosition: number) => {
-        const newState = [...state];
-        newState.splice(idPosition, 1);
-
-        return newState;
-    };
-
-    const removeAllowedUser = (state: typeof users, idPosition: number) => {
-        const newState = [...state];
-        const userProfile = newState[idPosition];
-
-        newState.splice(idPosition, 1, {
-            ...userProfile,
-            status: 'removed' as const,
-        });
-
-        return newState;
-    };
-
-    const restoreAllowedUser = (state: typeof users, idPosition: number) => {
-        const newState = [...state];
-        const userProfile = newState[idPosition];
-
-        newState.splice(idPosition, 1, {
-            ...userProfile,
-            status: 'idle' as const,
-        });
-
-        return newState;
-    };
-
-    const handleAllowedUsers = (
-        userId: typeof users[number]['id'],
-        status: 'added' | 'removed' | 'idle',
-    ) => {
-        setUsers((state) => {
-            const newState = [...state];
-            const usersIds = state.map(({ id }) => id);
-            const idPosition = usersIds.findIndex((id) => id === userId);
-
-            if (idPosition === -1) return [...state];
-
-            if (status === 'added')
-                return removeNewAllowedUser(newState, idPosition);
-            if (status === 'removed')
-                return restoreAllowedUser(newState, idPosition);
-            if (status === 'idle')
-                return removeAllowedUser(newState, idPosition);
-
-            return newState;
-        });
-    };
-
-    useEffect(() => {
-        if (!usersProfilesQuery.data) return;
-
-        const usersProfiles = usersProfilesQuery.data.map((userProfile) => {
-            const { image_url, ...rest } = userProfile;
-
-            return {
-                ...rest,
-                imageURL: image_url,
-                status:
-                    rest.id === ownerSpotifyId
-                        ? ('permanent' as const)
-                        : ('idle' as const),
-            };
-        });
-
-        setUsers(usersProfiles);
-    }, [usersProfilesQuery.data, ownerSpotifyId]);
+    const {
+        users,
+        addNewAllowedUser,
+        handleAllowedUsers,
+        handleSubmit,
+        isLoading,
+    } = useAllowedUsers({ playlist, allowedUsers, ownerSpotifyId });
 
     return (
         <div>
@@ -201,16 +86,19 @@ const Playlist: NextPage<PlaylistProps> = ({
                             {users.map((allowedUser) => {
                                 const { imageURL, name, id, status } =
                                     allowedUser;
+                                const imageSrc = imageURL || '/notDefined';
+
                                 return (
                                     <div
                                         key={id}
                                         style={{ border: 'solid white' }}
                                     >
-                                        <img
-                                            src={imageURL || '/notDefined'}
+                                        <Image
+                                            src={imageSrc}
                                             alt="logo"
                                             width="64"
                                             height="64"
+                                            loader={() => imageSrc}
                                         />
                                         {`${id} | ${name} | `}
                                         {status &&
@@ -244,23 +132,8 @@ const Playlist: NextPage<PlaylistProps> = ({
                                     </div>
                                 );
                             })}
-                            <button
-                                onClick={() =>
-                                    usersProfilesMutation.mutate(
-                                        users
-                                            .filter(
-                                                ({ status }) =>
-                                                    status !== 'removed',
-                                            )
-                                            .map(({ id }) => id),
-                                    )
-                                }
-                            >
-                                Save
-                            </button>
-                            {(usersProfilesMutation.isLoading ||
-                                usersProfilesQuery.isLoading) &&
-                                'Saving'}
+                            <button onClick={() => handleSubmit()}>Save</button>
+                            {isLoading && 'Saving'}
                         </div>
                         <UsersSearchBox
                             allowedUsersIds={users.map(({ id }) => id)}
